requirements.txt
asgiref==3.5.2
Django==4.0.8
sqlparse==0.4.3
tzdata==2022.7

djangorestframework~=3.13.1